implemented simple working 4 point anti aliasing

+++ Raytracer/src/camera.cpp
@@ -36,7 +36,7 @@ Ray PerspectiveCamera::ConstructRay(float pixelX, float pixelY) const
 	const float toRadians = (float)M_PI / 180.0f;
 	double Px = ToCanonicalSpace(pixelX, (float)m_resolutionX) * tan(m_degreesFOV / 2 * toRadians) * aspectRatio;
 	double Py = ToCanonicalSpace(pixelY, (float)m_resolutionY) * tan(m_degreesFOV / 2 * toRadians);
-	Vector3 rayDirection = Vector3(Px, Py, m_position.z + 1) - m_position;
+	Vector3 rayDirection = Vector3((float)Px, (float)Py, m_position.z + 1) - m_position;
 	rayDirection.Normalize();
 
 	return Ray(m_position, rayDirection);
